// frontend/src/api/generated/models/updateTaskRequestDoc.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * acpms-server
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTaskRequestDoc {
  /** @nullable */
  assigned_to?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  sprint_id?: string | null;
  status: string;
  /** @nullable */
  title?: string | null;
}